% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{detect_data_changes}
\alias{detect_data_changes}
\title{Detect if dataset has changed by comparing MD5 checksums}
\usage{
detect_data_changes(
  file_path,
  dataset_name,
  checksums_file = ".checksums.json"
)
}
\arguments{
\item{file_path}{Character. Path to the parquet file to check}

\item{dataset_name}{Character. Name of the dataset (for checksum lookup)}

\item{checksums_file}{Character. Path to the checksums JSON file.
Default is '.checksums.json'}
}
\value{
Logical. TRUE if data has changed (or is new), FALSE if unchanged
}
\description{
Detect if dataset has changed by comparing MD5 checksums
}
