% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_url}
\alias{get_url}
\title{Custom function to retrieve the codebook URL}
\usage{
get_url(table)
}
\arguments{
\item{table}{Character. The table where variable information is needed.}
}
\value{
Full URL to CDC data documentation, codebook, & frequencies.
}
\description{
Custom function to retrieve the codebook URL
}
