% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_dataset_config}
\alias{load_dataset_config}
\title{Load dataset configuration from YAML file}
\usage{
load_dataset_config(config_file = "inst/extdata/datasets.yml")
}
\arguments{
\item{config_file}{Character. Path to the datasets.yml file.
Default is 'inst/extdata/datasets.yml'}
}
\value{
Data frame with dataset configuration
}
\description{
Load dataset configuration from YAML file
}
