% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pull_nhanes}
\alias{pull_nhanes}
\title{Function to pull all datasets with the same base name}
\usage{
pull_nhanes(nhanes_table, selected_variables = NULL, save = TRUE)
}
\arguments{
\item{nhanes_table}{Character. NOT case-sensitive! Lowercase is allowed.}

\item{selected_variables}{Character vector. Default is NULL to grab all variables
from the dataset. Do not use \code{everything()}... just don't enter an argument value.}

\item{save}{Logical. Save as new .rda and .parquet files. Default is TRUE}
}
\value{
Tibble of datasets across cycles; using \code{bind_rows()}.
}
\description{
Function to pull all datasets with the same base name
}
