% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_r2}
\alias{read_r2}
\title{Read the dataset files from the public NHANES storage:}
\usage{
read_r2(dataset, type = "parquet")
}
\arguments{
\item{dataset}{Character. NHANES dataset base name. Not case sensitive.}

\item{type}{Character. Either "parquet" (faster) or "rda".}
}
\description{
Read the dataset files from the public NHANES storage:
}
\details{
This function downloads NHANES data file from nhanes.kylegrealis.com.
}
\examples{
trigly <- read_r2('trigly')
acq <- read_r2('acq')
}
