% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{term_search}
\alias{term_search}
\title{Custom wrapped function to nhanesA::nhanesSearch for concise output}
\usage{
term_search(var)
}
\arguments{
\item{var}{Character. Term or phrase to search}
}
\value{
Smaller output dataset
}
\description{
Custom wrapped function to nhanesA::nhanesSearch for concise output
}
