% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{update_checksum}
\alias{update_checksum}
\title{Update checksums file with new hash for a dataset}
\usage{
update_checksum(dataset_name, file_path, checksums_file = ".checksums.json")
}
\arguments{
\item{dataset_name}{Character. Name of the dataset}

\item{file_path}{Character. Path to the parquet file}

\item{checksums_file}{Character. Path to the checksums JSON file.
Default is '.checksums.json'}
}
\value{
NULL (writes to file)
}
\description{
Update checksums file with new hash for a dataset
}
