% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{var_search}
\alias{var_search}
\title{Custom wrapped function to nhanes::nhanesSearchVarName for concise output}
\usage{
var_search(var)
}
\arguments{
\item{var}{Character. Variable name}
}
\value{
Smaller output dataset
}
\description{
Custom wrapped function to nhanes::nhanesSearchVarName for concise output
}
